Migrating activity tests to androidTest type
